+++ packages/figma/src/entities/data/__generated__/component-sets/index.d.ts
@@ -1,5 +1,5 @@
-export declare const privateTemplateAttechmentField: {
-  "name": "privateTemplateAttechmentField",
+export declare const privateTemplateAttachmentField: {
+  "name": "privateTemplateAttachmentField",
   "key": "5ba20e248e9cd0292fc285488b2ed3b3145d37b0",
   "componentPropertyDefinitions": {
     "Show Header#40606:8": {
@@ -218,6 +218,20 @@ export declare const privateComponentFieldHeaderSuffix: {
         "Ghost Button",
         "Icon",
         "Custom"
+      ]
+    }
+  }
+};
+
+export declare const privateComponentIdentityPlaceholder: {
+  "name": "privateComponentIdentityPlaceholder",
+  "key": "7d7ae18cabac6ebad19934516de8e7e256854eba",
+  "componentPropertyDefinitions": {
+    "Identity": {
+      "type": "VARIANT",
+      "variantOptions": [
+        "Person",
+        "Business"
       ]
     }
   }
@@ -2079,11 +2093,11 @@ export declare const componentBottomSheet: {
     "Header Layout": {
       "type": "VARIANT",
       "variantOptions": [
-        "Bottom Left",
-        "None",
-        "Bottom Center",
+        "Top Left",
         "Top Center",
-        "Top Left"
+        "Bottom Left",
+        "Bottom Center",
+        "None"
       ]
     }
   }
@@ -2648,7 +2662,7 @@ export declare const componentImageFrame: {
         "80",
         "96",
         "120",
-        "\bFree"
+        "Free"
       ]
     },
     "Rounded": {
@@ -3541,8 +3555,7 @@ export declare const componentSwitch: {
       "type": "VARIANT",
       "variantOptions": [
         "Label Last",
-        "Label First",
-        "🚫[Switch Mark 사용] Switch Only"
+        "Label First"
       ]
     }
   }