+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @seed-design/mcp
 
+## 1.3.5
+
+### Patch Changes
+
+- Updated dependencies [d6df976]
+  - @seed-design/figma@1.3.5
+
+## 1.3.4
+
+### Patch Changes
+
+- Updated dependencies [934a877]
+  - @seed-design/figma@1.3.4
+
 ## 1.3.3
 
 ### Patch Changes